k8s_sre_env/server/requirements.txt
openenv-core==0.2.1
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0
kubernetes>=28.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
requests>=2.25.0
huggingface_hub>=0.20.0
openai>=1.0.0
